ENH: Offset line starts when there is no source separation
If lines are all starting at the same point, we can just offset
the lines and not bother with the padding, which generally works
better.
So we do this, with a new input argument to control how much space
to use. The line backing now defaults to `false` if there is no
source separation.