ENH: Add legacy support for R2007a-2013b inputParser
Changed parser handling to use either
- `parser.addParameter` in >R2014a
- `parser.addParamValue` in R2007a-2013b

In R2014a, the method used to add parameters to inputParser was
changed to `addParameter` and `addParamValue` was deprecated.

From an external point-of-view, the two appear to be interchangable,
so the solution of using `addParameter` if it is present and if not
passing exactly the same inputs to `addParamValue` seems to work
fine.

This closes issue #11.