ENH: Don't use LaTeX mode for p text; use char(8727) star
Change the default star icon to be the character `char(8727)`,
which is the asterisk operator, unicode code point U+2217. This
is what LaTeX was using when it was given the `'*'` input
previously.

Specifying the unicode character we really want directly means we
no longer need to use LaTeX mode for the text (in fact it does not
work when using `char(8727)` input), so this has been dropped. This
will hopefully result in a superior user experience, since the text
will behaviour like the default settings instead.