BUG: Avoid narginchk on versions <R2011b
Before release R2011b, narginchk did not exist and you were
supposed to do `error(nargchk())`. Instead of using nargchk, we
just checking the number of arguments manually when narginchk
is not present. This means we can return the same error identifier
and message as if narginchk was present. Also it avoids having an
additional case to work with (what if narginchk and nargchk are
both missing).